set-status/requirements.txt
googlesearch-python
requests
python-dotenv
customtkinter
py2app